move paramters and state to separate file
moved here from admm_optimizer.py
