+++ qiskit/optimization/algorithms/admm/admm_optimizer.py
@@ -20,136 +20,18 @@ from typing import List, Optional, Any, Tuple
 
 import numpy as np
 from cplex import SparsePair
-from .cplex_optimizer import CplexOptimizer
-from .optimization_algorithm import OptimizationAlgorithm
-from ..problems.optimization_problem import OptimizationProblem
-from ..problems.variables import CPX_BINARY, CPX_CONTINUOUS
-from ..results.optimization_result import OptimizationResult
-from ..utils.qiskit_optimization_error import QiskitOptimizationError
+from .admm_state import ADMMState
+from .admm_parameters import ADMMParameters, UPDATE_RHO_BY_TEN_PERCENT
+from ..cplex_optimizer import CplexOptimizer
+from ..optimization_algorithm import OptimizationAlgorithm
+from ...converters.optimization_problem_to_admm_state import OptimizationProblemToADMMState
+from ...problems.optimization_problem import OptimizationProblem
+from ...problems.variables import CPX_BINARY, CPX_CONTINUOUS
+from ...results.optimization_result import OptimizationResult
 
-UPDATE_RHO_BY_TEN_PERCENT = 0
 UPDATE_RHO_BY_RESIDUALS = 1
 
 logger = logging.getLogger(__name__)
-
-
-class ADMMParameters:
-    """Defines a set of parameters for ADMM optimizer."""
-
-    def __init__(self, rho_initial: float = 10000, factor_c: float = 100000, beta: float = 1000,
-                 max_iter: int = 10, tol: float = 1.e-4, max_time: float = np.inf,
-                 three_block: bool = True, vary_rho: int = UPDATE_RHO_BY_TEN_PERCENT,
-                 tau_incr: float = 2, tau_decr: float = 2, mu_res: float = 10,
-                 mu_merit: float = 1000) -> None:
-        """Defines parameters for ADMM optimizer and their default values.
-
-        Args:
-            rho_initial: Initial value of rho parameter of ADMM.
-            factor_c: Penalizing factor for equality constraints, when mapping to QUBO.
-            beta: Penalization for y decision variables.
-            max_iter: Maximum number of iterations for ADMM.
-            tol: Tolerance for the residual convergence.
-            max_time: Maximum running time (in seconds) for ADMM.
-            three_block: Boolean flag to select the 3-block ADMM implementation.
-            vary_rho: Flag to select the rule to update rho.
-                If set to 0, then rho increases by 10% at each iteration.
-                If set to 1, then rho is modified according to primal and dual residuals.
-            tau_incr: Parameter used in the rho update (UPDATE_RHO_BY_RESIDUALS).
-                The update rule can be found in:
-                Boyd, S., Parikh, N., Chu, E., Peleato, B., & Eckstein, J. (2011).
-                Distributed optimization and statistical learning via the alternating
-                direction method of multipliers.
-                Foundations and Trends® in Machine learning, 3(1), 1-122.
-            tau_decr: Parameter used in the rho update (UPDATE_RHO_BY_RESIDUALS).
-            mu_res: Parameter used in the rho update (UPDATE_RHO_BY_RESIDUALS).
-            mu_merit: Penalization for constraint residual. Used to compute the merit values.
-        """
-        super().__init__()
-        self.mu_merit = mu_merit
-        self.mu_res = mu_res
-        self.tau_decr = tau_decr
-        self.tau_incr = tau_incr
-        self.vary_rho = vary_rho
-        self.three_block = three_block
-        self.max_time = max_time
-        self.tol = tol
-        self.max_iter = max_iter
-        self.factor_c = factor_c
-        self.beta = beta
-        self.rho_initial = rho_initial
-
-
-class ADMMState:
-    """Internal computation state of the ADMM implementation.
-
-    The state keeps track of various variables are stored that are being updated during problem
-    solving. The values are relevant to the problem being solved. The state is recreated for each
-    optimization problem. State is returned as the third value.
-    """
-
-    def __init__(self,
-                 op: OptimizationProblem,
-                 binary_indices: List[int],
-                 continuous_indices: List[int],
-                 rho_initial: float) -> None:
-        """Constructs an internal computation state of the ADMM implementation.
-
-        Args:
-            op: The optimization problem being solved.
-            binary_indices: Indices of the binary decision variables of the original problem.
-            continuous_indices: Indices of the continuous decision variables of the original
-             problem.
-            rho_initial: Initial value of the rho parameter.
-        """
-        super().__init__()
-
-        # Optimization problem itself
-        self.op = op
-        # Indices of the variables
-        self.binary_indices = binary_indices
-        self.continuous_indices = continuous_indices
-        self.sense = op.objective.get_sense()
-
-        # define heavily used matrix, they are used at each iteration, so let's cache them,
-        # they are np.ndarrays
-        # pylint:disable=invalid-name
-        # objective
-        self.q0 = None
-        self.c0 = None
-        self.q1 = None
-        self.c1 = None
-        # constraints
-        self.a0 = None
-        self.b0 = None
-        self.a1 = None
-        self.b1 = None
-        self.a2 = None
-        self.a3 = None
-        self.b2 = None
-        self.a4 = None
-        self.b3 = None
-
-        # These are the parameters that are updated in the ADMM iterations.
-        self.u: np.ndarray = np.zeros(len(continuous_indices))
-        binary_size = len(binary_indices)
-        self.x0: np.ndarray = np.zeros(binary_size)
-        self.z: np.ndarray = np.zeros(binary_size)
-        self.z_init: np.ndarray = self.z
-        self.y: np.ndarray = np.zeros(binary_size)
-        self.lambda_mult: np.ndarray = np.zeros(binary_size)
-
-        # The following structures store quantities obtained in each ADMM iteration.
-        self.cost_iterates = []
-        self.residuals = []
-        self.dual_residuals = []
-        self.cons_r = []
-        self.merits = []
-        self.lambdas = []
-        self.x0_saved = []
-        self.u_saved = []
-        self.z_saved = []
-        self.y_saved = []
-        self.rho = rho_initial
 
 
 class ADMMOptimizerResult(OptimizationResult):
@@ -261,17 +143,8 @@ class ADMMOptimizer(OptimizationAlgorithm):
         Raises:
             QiskitOptimizationError: If the problem is incompatible with the optimizer.
         """
-        # parse problem and convert to an ADMM specific representation.
-        binary_indices = self._get_variable_indices(problem, CPX_BINARY)
-        continuous_indices = self._get_variable_indices(problem, CPX_CONTINUOUS)
-
-        # create our computation state.
-        self._state = ADMMState(problem, binary_indices,
-                                continuous_indices, self._params.rho_initial)
-
-        # convert optimization problem to a set of matrices and vector that are used
-        # at each iteration.
-        self._convert_problem_representation()
+        # parse problem and convert to an ADMM specific representation
+        self._state = OptimizationProblemToADMMState(self._params).encode(problem)
 
         start_time = time.time()
         # we have not stated our computations yet, so elapsed time initialized as zero.
@@ -281,7 +154,7 @@ class ADMMOptimizer(OptimizationAlgorithm):
 
         while iteration < self._params.max_iter and residual > self._params.tol \
                 and elapsed_time < self._params.max_time:
-            if binary_indices:
+            if self._state.binary_indices:
                 op1 = self._create_step1_problem()
                 self._state.x0 = self._update_x0(op1)
             # else, no binary variables exist,
@@ -296,7 +169,7 @@ class ADMMOptimizer(OptimizationAlgorithm):
             logger.debug("z=%s", self._state.z)
 
             if self._params.three_block:
-                if binary_indices:
+                if self._state.binary_indices:
                     op3 = self._create_step3_problem()
                     self._state.y = self._update_y(op3)
                 # debug
@@ -376,234 +249,6 @@ class ADMMOptimizer(OptimizationAlgorithm):
         for i, continuous_index in enumerate(self._state.continuous_indices):
             solution[continuous_index] = continuous_solutions[i]
         return solution
-
-    def _convert_problem_representation(self) -> None:
-        """Converts problem representation into set of matrices and vectors.
-
-        Specifically, the optimization problem is represented as:
-
-        min_{x0, u} x0^T q0 x0 + c0^T x0 + u^T q1 u + c1^T u
-
-        s.t. a0 x0 = b0
-            a1 x0 <= b1
-            a2 z + a3 u <= b2
-            a4 u <= b3
-        """
-        # objective
-        self._state.q0 = self._get_q(self._state.binary_indices)
-        self._state.c0 = self._get_c(self._state.binary_indices)
-        self._state.q1 = self._get_q(self._state.continuous_indices)
-        self._state.c1 = self._get_c(self._state.continuous_indices)
-        # constraints
-        self._state.a0, self._state.b0 = self._get_a0_b0()
-        self._state.a1, self._state.b1 = self._get_a1_b1()
-        self._state.a2, self._state.a3, self._state.b2 = self._get_a2_a3_b2()
-        self._state.a4, self._state.b3 = self._get_a4_b3()
-
-    def _get_q(self, variable_indices: List[int]) -> np.ndarray:
-        """Constructs a quadratic matrix for the variables with the specified indices
-        from the quadratic terms in the objective.
-
-        Args:
-            variable_indices: Variable indices to look for.
-
-        Returns:
-            A matrix as a numpy array of the shape(len(variable_indices), len(variable_indices)).
-        """
-        size = len(variable_indices)
-        q = np.zeros(shape=(size, size))
-        # fill in the matrix
-        # in fact we use re-indexed variables
-        for i, var_index_i in enumerate(variable_indices):
-            for j, var_index_j in enumerate(variable_indices):
-                q[i, j] = self._state.op.objective.get_quadratic_coefficients(var_index_i,
-                                                                              var_index_j)
-
-        # flip the sign, according to the optimization sense, e.g. sense == 1 if minimize,
-        # sense == -1 if maximize.
-        return q * self._state.sense
-
-    def _get_c(self, variable_indices: List[int]) -> np.ndarray:
-        """Constructs a vector for the variables with the specified indices from the linear terms
-        in the objective.
-
-        Args:
-            variable_indices: Variable indices to look for.
-
-        Returns:
-            A numpy array of the shape(len(variable_indices)).
-        """
-        c = np.array(self._state.op.objective.get_linear(variable_indices))
-        # flip the sign, according to the optimization sense, e.g. sense == 1 if minimize,
-        # sense == -1 if maximize.
-        c *= self._state.sense
-        return c
-
-    def _assign_row_values(self, matrix: List[List[float]], vector: List[float],
-                           constraint_index: int, variable_indices: List[int]) -> None:
-        """Appends a row to the specified matrix and vector based on the constraint specified by
-        the index using specified variables.
-
-        Args:
-            matrix: A matrix to extend.
-            vector: A vector to expand.
-            constraint_index: Constraint index to look for.
-            variable_indices: Variables to look for.
-        """
-        # assign matrix row.
-        row = []
-        for var_index in variable_indices:
-            row.append(self._state.op.linear_constraints.get_coefficients(constraint_index,
-                                                                          var_index))
-        matrix.append(row)
-
-        # assign vector row.
-        vector.append(self._state.op.linear_constraints.get_rhs(constraint_index))
-
-        # flip the sign if constraint is G, we want L constraints.
-        if self._state.op.linear_constraints.get_senses(constraint_index) == "G":
-            # invert the sign to make constraint "L".
-            matrix[-1] = [-1 * el for el in matrix[-1]]
-            vector[-1] = -1 * vector[-1]
-
-    @staticmethod
-    def _create_ndarrays(matrix: List[List[float]], vector: List[float], size: int) \
-            -> Tuple[np.ndarray, np.ndarray]:
-        """Converts representation of a matrix and a vector in form of lists to numpy array.
-
-        Args:
-            matrix: matrix to convert.
-            vector: vector to convert.
-            size: size to create matrix and vector.
-
-        Returns:
-            Converted matrix and vector as numpy arrays.
-        """
-        # if we don't have such constraints, return just dummy arrays.
-        if len(matrix) != 0:
-            return np.array(matrix), np.array(vector)
-        else:
-            return np.array([0] * size).reshape((1, -1)), np.zeros(shape=(1,))
-
-    def _get_a0_b0(self) -> Tuple[np.ndarray, np.ndarray]:
-        """Constructs a matrix and a vector from the constraints in a form of Ax = b, where
-        x is a vector of binary variables.
-
-        Returns:
-            Corresponding matrix and vector as numpy arrays.
-
-        Raises:
-            ValueError: if the problem is not suitable for this optimizer.
-        """
-        matrix = []
-        vector = []
-
-        senses = self._state.op.linear_constraints.get_senses()
-        index_set = set(self._state.binary_indices)
-        for constraint_index, sense in enumerate(senses):
-            # we check only equality constraints here.
-            if sense != "E":
-                continue
-            row = self._state.op.linear_constraints.get_rows(constraint_index)
-            if set(row.ind).issubset(index_set):
-                self._assign_row_values(matrix, vector,
-                                        constraint_index, self._state.binary_indices)
-            else:
-                raise ValueError(
-                    "Linear constraint with the 'E' sense must contain only binary variables, "
-                    "row indices: {}, binary variable indices: {}".format(
-                        row, self._state.binary_indices))
-
-        return self._create_ndarrays(matrix, vector, len(self._state.binary_indices))
-
-    def _get_inequality_matrix_and_vector(self, variable_indices: List[int]) \
-            -> Tuple[List[List[float]], List[float]]:
-        """Constructs a matrix and a vector from the constraints in a form of Ax <= b, where
-        x is a vector of variables specified by the indices.
-
-        Args:
-            variable_indices: Variable indices to look for.
-
-        Returns:
-            A list based representation of the matrix and the vector.
-
-        Raises:
-            QiskitOptimizationError: If contraint types "E" or "R" are used, which are not yet
-                supported.
-        """
-        matrix = []
-        vector = []
-        senses = self._state.op.linear_constraints.get_senses()
-
-        index_set = set(variable_indices)
-        for constraint_index, sense in enumerate(senses):
-            if sense in ("E", "R"):
-                # TODO: Ranged constraints should be supported
-                raise QiskitOptimizationError('Ranged constraints (E,R) are not yet supported.')
-
-            # sense either G or L.
-            row = self._state.op.linear_constraints.get_rows(constraint_index)
-            if set(row.ind).issubset(index_set):
-                self._assign_row_values(matrix, vector, constraint_index, variable_indices)
-
-        return matrix, vector
-
-    def _get_a1_b1(self) -> (np.ndarray, np.ndarray):
-        """Constructs a matrix and a vector from the constraints in a form of Ax <= b, where
-        x is a vector of binary variables.
-
-        Returns:
-            A numpy based representation of the matrix and the vector.
-        """
-        matrix, vector = self._get_inequality_matrix_and_vector(self._state.binary_indices)
-        return self._create_ndarrays(matrix, vector, len(self._state.binary_indices))
-
-    def _get_a4_b3(self) -> (np.ndarray, np.ndarray):
-        """Constructs a matrix and a vector from the constraints in a form of Au <= b, where
-        u is a vector of continuous variables.
-
-        Returns:
-            A numpy based representation of the matrix and the vector.
-        """
-        matrix, vector = self._get_inequality_matrix_and_vector(self._state.continuous_indices)
-        return self._create_ndarrays(matrix, vector, len(self._state.continuous_indices))
-
-    def _get_a2_a3_b2(self) -> (np.ndarray, np.ndarray, np.ndarray):
-        """Constructs matrices and a vector from the constraints in a form of A_2x + A_3u <= b,
-        where x is a vector of binary variables and u is a vector of continuous variables.
-
-        Returns:
-            A numpy representation of two matrices and one vector.
-
-        Raises:
-            QiskitOptimizationError: If contraint types "E" or "R" are used, which are not yet
-                supported.
-        """
-        matrix = []
-        vector = []
-        senses = self._state.op.linear_constraints.get_senses()
-
-        binary_index_set = set(self._state.binary_indices)
-        continuous_index_set = set(self._state.continuous_indices)
-        all_variables = self._state.binary_indices + self._state.continuous_indices
-        for constraint_index, sense in enumerate(senses):
-            if sense in ("E", "R"):
-                # TODO: Ranged constraints should be supported as well
-                raise QiskitOptimizationError('Ranged constraints (E,R) are not yet supported.')
-
-            # sense either G or L.
-            row = self._state.op.linear_constraints.get_rows(constraint_index)
-            row_indices = set(row.ind)
-            # we must have a least one binary and one continuous variable,
-            # otherwise it is another type of constraints.
-            if len(row_indices & binary_index_set) != 0 and len(
-                    row_indices & continuous_index_set) != 0:
-                self._assign_row_values(matrix, vector, constraint_index, all_variables)
-
-        matrix, b_2 = self._create_ndarrays(matrix, vector, len(all_variables))
-        a_2 = matrix[:, 0:len(self._state.binary_indices)]
-        a_3 = matrix[:, len(self._state.binary_indices):]
-        return a_2, a_3, b_2
 
     def _create_step1_problem(self) -> OptimizationProblem:
         """Creates a step 1 sub-problem.
